Add support for including additional artifacts in Docker Image
The Waypoint team has a requirement to include an entrypoint binary that is built
alongside their product binary in the Docker image they build. This adds support
for including arbitrary content in the workspace when building a Docker image with
the following requirements:

* The artifact must be a zip file
* The artifact must be available in workflow uploaded artifacts store

Artifacts are downloaded to a subdirectory named `additional` under the specified
`zip_location`. This avoids name conflicts with the primary zip since we must download
all uploaded artifacts in this step. The contents are extracted to the same directory
as the primary zip. It is left to the user to ensure there are no name conflicts
between the extracted files.

Signed-off-by: Scott Macfarlane <smacfarlane@hashicorp.com>